requirements.txt
cryptocompare>=0.7.6
python-dotenv>=1.0.0
plotly>=5.18.0
pandas>=2.1.0
kaleido>=0.2.1
